fix(EmailController): Corrige erro de índice indefinido ao exibir nome do mês para frequência
O erro 'Undefined index: 05' ocorria na view `emails.indexAttendanceRecords`
porque a variável `$month` (ex: '05' para Maio) era usada como string para
acessar o array `$meses`, que esperava uma chave inteira (ex: 5).

A correção foi feita no método `indexAttendanceRecords` do `EmailController`,
garantindo que `$month` seja convertido para inteiro (`(int)$month`) antes de
ser passado para a view e utilizado na consulta de frequências.

Isso previne o erro e assegura a exibição correta do nome do mês.

+++ app/Http/Controllers/TutorController.php
@@ -81,7 +81,7 @@ class TutorController extends Controller
         Frequency::where('student_id', $selection->student_id)
             ->where('school_class_id', $selection->school_class_id)
             ->where('registered', false)
-            ->where('month', '>=', date('m'))
+            ->where('month', '>=', (int)date('m'))
             ->delete();
 
         $selection->sitatl = "Desligado";
